bnxt_re/lib: Direct Verbs: Export HSI file
Extract HSI definitions (WQE/CQE structures and enums) from
bnxt_re-abi.h into a standalone bnxt_re-hsi.h and add it to
publish_headers() so it is installed under infiniband/.

This will make HSI definitions available to DV applications,
that implement their own datapath.
